+++ pkg/cmd/variable_test.go
@@ -12,6 +12,7 @@ type mockVariableClient struct {
 	t              *testing.T
 	get            func(*testing.T, string) ([]byte, error)
 	getWithVersion func(*testing.T, string, int) ([]byte, error)
+	getBatch       func(*testing.T, []string) (map[string][]byte, error)
 	set            func(*testing.T, string, string) error
 }
 
@@ -20,6 +21,9 @@ func (m mockVariableClient) RetrieveSecret(path string) ([]byte, error) {
 }
 func (m mockVariableClient) RetrieveSecretWithVersion(path string, version int) ([]byte, error) {
 	return m.getWithVersion(m.t, path, version)
+}
+func (m mockVariableClient) RetrieveBatchSecretsSafe(paths []string) (map[string][]byte, error) {
+	return m.getBatch(m.t, paths)
 }
 func (m mockVariableClient) AddSecret(path string, value string) error {
 	return m.set(m.t, path, value)
@@ -30,6 +34,7 @@ var variableCmdTestCases = []struct {
 	args               []string
 	get                func(t *testing.T, path string) ([]byte, error)
 	getWithVersion     func(t *testing.T, path string, version int) ([]byte, error)
+	getBatch           func(t *testing.T, paths []string) (map[string][]byte, error)
 	set                func(t *testing.T, path string, value string) error
 	clientFactoryError error
 	assert             func(t *testing.T, stdout string, stderr string, err error)
@@ -58,11 +63,11 @@ var variableCmdTestCases = []struct {
 	{
 		name: "get subcommand",
 		args: []string{"variable", "get", "-i", "meow"},
-		get: func(t *testing.T, path string) ([]byte, error) {
+		getBatch: func(t *testing.T, path []string) (map[string][]byte, error) {
 			// Assert on arguments
-			assert.Equal(t, "meow", path)
+			assert.Equal(t, ("meow"), path[0])
 
-			return []byte("moo"), nil
+			return map[string][]byte{"meow": []byte("moo")}, nil
 		},
 		assert: func(t *testing.T, stdout, stderr string, err error) {
 			assert.Contains(t, stdout, "moo")
@@ -71,7 +76,7 @@ var variableCmdTestCases = []struct {
 	{
 		name: "get subcommand error",
 		args: []string{"variable", "get", "-i", "meow"},
-		get: func(t *testing.T, path string) ([]byte, error) {
+		getBatch: func(t *testing.T, path []string) (map[string][]byte, error) {
 			return nil, fmt.Errorf("%s", "get error")
 		},
 		assert: func(t *testing.T, stdout, stderr string, err error) {
@@ -104,6 +109,27 @@ var variableCmdTestCases = []struct {
 		clientFactoryError: fmt.Errorf("%s", "client factory error"),
 		assert: func(t *testing.T, stdout, stderr string, err error) {
 			assert.Contains(t, stderr, "Error: client factory error\n")
+		},
+	},
+	{
+		name: "get two variables",
+		args: []string{"variable", "get", "-i", "meow,woof"},
+		getBatch: func(t *testing.T, path []string) (map[string][]byte, error) {
+			// Assert on arguments
+			assert.Equal(t, "meow", path[0])
+			assert.Equal(t, "woof", path[1])
+			return map[string][]byte{"meow": []byte("moo"), "woof": []byte("quack")}, nil
+		},
+		assert: func(t *testing.T, stdout, stderr string, err error) {
+			assert.Contains(t, stdout, "moo")
+			assert.Contains(t, stdout, "quack")
+		},
+	},
+	{
+		name: "get two variables with version",
+		args: []string{"variable", "get", "-i", "meow,woof", "-v", "1"},
+		assert: func(t *testing.T, stdout, stderr string, err error) {
+			assert.Contains(t, stderr, "version can not be used with multiple variables")
 		},
 	},
 	{
@@ -153,7 +179,7 @@ func TestVariableCmd(t *testing.T) {
 	for _, tc := range variableCmdTestCases {
 		t.Run(tc.name, func(t *testing.T) {
 			mockClient := mockVariableClient{
-				t: t, set: tc.set, get: tc.get, getWithVersion: tc.getWithVersion,
+				t: t, set: tc.set, get: tc.get, getWithVersion: tc.getWithVersion, getBatch: tc.getBatch,
 			}
 
 			cmd := newVariableCmd(